+++ example/lib/src/config_example/home_screen.dart
@@ -28,7 +28,7 @@ class _MyHomePageState extends State<MyHomePage> {
   void _incrementCounter() {
     setState(() {
       _counter +=
-          Config.of<GeneralConfig>(context).get(GeneralConfig.counterScaler);
+          Config.from<GeneralConfig>(context).get(GeneralConfig.counterScaler);
     });
   }
 
@@ -36,7 +36,7 @@ class _MyHomePageState extends State<MyHomePage> {
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
-        backgroundColor: Theme.of(context).colorScheme.inversePrimary,
+        backgroundColor: Theme.of(context).colorScheme.background,
         title: Text(widget.title),
         actions: [
           IconButton(