+++ example/lib/src/multi_settings_app/app.dart
@@ -1,9 +1,8 @@
 import 'package:example/src/multi_settings_app/home_screen.dart';
 import 'package:example/src/multi_settings_app/settings.dart';
+import 'package:example/src/simple_app/settings.dart';
 import 'package:flutter/material.dart';
 import 'package:settings_provider/settings_provider.dart';
-
-import '../../multi_main.dart';
 
 class MultiSettingsApp extends StatelessWidget {
   const MultiSettingsApp({super.key});
@@ -13,7 +12,9 @@ class MultiSettingsApp extends StatelessWidget {
     return Builder(builder: (context) {
       ThemeData theme;
       // uses setting to get data from SettingsData
-      if (context.listenSetting<HomeScreenSettings>().get(isDarkMode)) {
+      if (context
+          .listenSetting<HomeScreenSettings>()
+          .get(GeneralSettings.isDarkMode)) {
         theme = ThemeData.dark(useMaterial3: true);
       } else {
         theme = ThemeData(useMaterial3: true);