+++ example/lib/src/multi_settings_app/settings.dart
@@ -1,16 +1,31 @@
 import 'package:settings_provider/settings_provider.dart';
 
-const Property<bool> isDarkMode = Property(
-  defaultValue: false,
-  id: 'isDarkMode',
-  isLocalStored: true,
-);
+// A separate model for home screen settings
+class HomeScreenSettings extends SettingsModel {
+  @override
+  List<Property> get properties => [isDarkMode, counterScaler];
 
-const Property<int> counterScaler =
-    Property(defaultValue: 1, id: 'counterScaler');
+  @override
+  List<Scenario<Enum>>? get scenarios => null;
 
-List<Property> settings = [isDarkMode, counterScaler];
+  static const Property<bool> isDarkMode = Property(
+    defaultValue: false,
+    id: 'isDarkMode',
+    isLocalStored: true,
+  );
 
-const Property<String> name = Property(defaultValue: 'John', id: 'name');
+  static const Property<int> counterScaler =
+      Property(defaultValue: 1, id: 'counterScaler');
+}
 
-List<Property> personalSettings = [name];
+// A separate model for personal screen settings
+class PersonalScreenSettings extends SettingsModel {
+  @override
+  List<Property> get properties => [name];
+
+  @override
+  List<Scenario<Enum>>? get scenarios => null;
+
+  static const Property<String> name =
+      Property(defaultValue: 'John', id: 'name');
+}