+++ example/lib/src/simple_app/app.dart
@@ -13,7 +13,9 @@ class SimpleApp extends StatelessWidget {
     return Builder(builder: (context) {
       ThemeData theme;
       // uses setting to get data from SettingsData
-      if (Settings.of(context, listen: true).get(isDarkMode)) {
+      if (context
+          .listenSetting<GeneralSettings>()
+          .get(GeneralSettings.isDarkMode)) {
         theme = ThemeData.dark(useMaterial3: true);
       } else {
         theme = ThemeData(useMaterial3: true);