+++ example/lib/src/simple_app/settings.dart
@@ -1,12 +1,18 @@
 import 'package:settings_provider/settings_provider.dart';
 
-const isDarkMode = Property<bool>(
-  defaultValue: false,
-  id: 'isDarkMode',
-  isLocalStored: true,
-);
+class GeneralSettings extends SettingsModel {
+  @override
+  List<Property> get properties => [isDarkMode, counterScaler];
 
-const counterScaler =
-    Property<int>(defaultValue: 1, id: 'counterScaler', isLocalStored: true);
+  @override
+  List<Scenario<Enum>>? get scenarios => null;
 
-List<Property> settings = [isDarkMode, counterScaler];
+  static const Property<bool> isDarkMode = Property(
+    defaultValue: false,
+    id: 'isDarkMode',
+    isLocalStored: true,
+  );
+
+  static const Property<int> counterScaler =
+      Property(defaultValue: 1, id: 'counterScaler', isLocalStored: true);
+}