+++ example/lib/src/simple_app/settings_screen.dart
@@ -40,7 +40,10 @@ class _CounterScalerSettingsState extends State<CounterScalerSettings> {
   void initState() {
     super.initState();
     _textController = TextEditingController();
-    _textController.text = context.setting().get(counterScaler).toString();
+    _textController.text = context
+        .setting<GeneralSettings>()
+        .get(GeneralSettings.counterScaler)
+        .toString();
   }
 
   @override
@@ -67,11 +70,13 @@ class _CounterScalerSettingsState extends State<CounterScalerSettings> {
               controller: _textController,
               onTapOutside: (event) {
                 if (_scaler != 0) {
-                  context
-                      .setting()
-                      .update(counterScaler.copyWith(defaultValue: _scaler));
+                  context.setting<GeneralSettings>().update(GeneralSettings
+                      .counterScaler
+                      .copyWith(defaultValue: _scaler));
                 } else {
-                  context.setting().update(counterScaler);
+                  context
+                      .setting<GeneralSettings>()
+                      .update(GeneralSettings.counterScaler);
                 }
               },
               onChanged: (value) async {
@@ -88,8 +93,8 @@ class _CounterScalerSettingsState extends State<CounterScalerSettings> {
                           TextButton(
                               onPressed: () {
                                 value = context
-                                    .setting()
-                                    .get(counterScaler)
+                                    .setting<GeneralSettings>()
+                                    .get(GeneralSettings.counterScaler)
                                     .toString();
                                 Navigator.pop(context);
                               },
@@ -127,13 +132,17 @@ class DarkModeSetting extends StatelessWidget {
         ),
         Switch(
           // switch use the getSetting() method to get the value
-          value: context.listenSetting().get(isDarkMode),
+          value: context
+              .listenSetting<GeneralSettings>()
+              .get(GeneralSettings.isDarkMode),
           onChanged: (newValue) {
-            if (newValue != context.setting().get(isDarkMode)) {
+            if (newValue !=
+                context
+                    .setting<GeneralSettings>()
+                    .get(GeneralSettings.isDarkMode)) {
               // if newValue is different we can update our settings
-              context
-                  .setting()
-                  .update(isDarkMode.copyWith(defaultValue: newValue));
+              context.setting<GeneralSettings>().update(
+                  GeneralSettings.isDarkMode.copyWith(defaultValue: newValue));
             }
           },
         ),