+++ lib/src/config/config_provider.dart
@@ -45,23 +45,59 @@ class Config<T extends SingleChildWidget> extends StatefulWidget {
     return initedProviders;
   }
 
+  /// The static `of()` method allows you to look up the implemented SettingsNotifier
+  /// that stores the controller to interact with settings.
+  ///
+  /// The `listen = true` option allows you to subscribe to changes in settings.
+  /// But we should use the method with this parameter only in the context of the widget tree
+  /// or in the `didChangeDependencies()` method.
+  @Deprecated(
+      'This method was split into two separate methods. For listening: `listenFrom()`, and for unsubscribed access `from()`.')
   static T of<T extends BaseSettingsModel>(BuildContext context,
       {bool listen = false}) {
     if (listen == true) {
       SettingsNotifier<T>? provider =
           context.dependOnInheritedWidgetOfExactType<SettingsNotifier<T>>();
       if (provider == null && null is! T) {
-        throw Exception('Not founded provided Config');
+        throw Exception('Not founded provided Settings');
       }
       return provider?.notifier as T;
     } else {
       SettingsNotifier<T>? provider =
           context.getInheritedWidgetOfExactType<SettingsNotifier<T>>();
       if (provider == null && null is! T) {
-        throw Exception('Not founded provided Config');
+        throw Exception('Not founded provided Settings');
       }
       return provider?.notifier as T;
     }
+  }
+
+  /// The static `listenFrom()` method allows you to look up the implemented SettingsNotifier
+  /// that stores the controller to interact with settings.
+  ///
+  /// This method allows you to subscribe to changes in settings.
+  /// But we should use the method with this parameter only in the context of the widget tree
+  /// or in the `didChangeDependencies()` method.
+  static T listenFrom<T extends BaseSettingsModel>(BuildContext context) {
+    SettingsNotifier<T>? provider =
+        context.dependOnInheritedWidgetOfExactType<SettingsNotifier<T>>();
+    if (provider == null && null is! T) {
+      throw Exception('Not founded provided Settings');
+    }
+    return provider?.notifier as T;
+  }
+
+  /// The static `from()` method allows you to look up the implemented SettingsNotifier
+  /// that stores the controller to interact with settings.
+  ///
+  /// This method allows you to get setting from settings.
+  static T from<T extends BaseSettingsModel>(BuildContext context) {
+    SettingsNotifier<T>? provider =
+        context.getInheritedWidgetOfExactType<SettingsNotifier<T>>();
+    if (provider == null && null is! T) {
+      throw Exception('Not founded provided Settings');
+    }
+    return provider?.notifier as T;
   }
 
   final Widget child;