+++ CHANGELOG.md
@@ -24,3 +24,9 @@
 * Improved interface of context extensions;
 * Generalized interface for configurations and settings separately;
 * Some improvements.
+
+## 0.2.1
+
+* Changes for standardizing access to settings. Now, `BuildContext()` extensions have limited options for accessing settings. The available context methods include `setting<T>()` and `listenSetting<T>()`. Additionally, the `of()` method has been deprecated in favor of two separate methods, `from()` and `listenFrom()`.
+* Now, to create settings, it is mandatory to create a class based on `SettingsModel()` for settings or a class based on `ConfigModel()` for configs. The type of this class must be specified when accessing settings through context methods.
+* Thanks to standardization, `ScenarioBuilder()` can now be used in both configs and regular settings.